# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synque Express Ticketing API",
    "version": "0.1.0",
    "description": "Complete ticketing and event management API with user profile management, promotions, and cancellation support"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.ticketing.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "Commerce"
        ],
        "summary": "Create new order",
        "description": "Create a new order for one or more products",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "items",
                  "channel_id",
                  "out_trade_no"
                ],
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "description": "ID of the user placing the order"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "qty": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  },
                  "channel_id": {
                    "type": "integer",
                    "description": "Sales channel identifier"
                  },
                  "out_trade_no": {
                    "type": "string",
                    "description": "External transaction number for idempotency"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order_id": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "paid",
                        "cancelled"
                      ]
                    },
                    "amounts": {
                      "type": "object",
                      "properties": {
                        "subtotal": {
                          "type": "number",
                          "format": "decimal"
                        },
                        "discount": {
                          "type": "number",
                          "format": "decimal"
                        },
                        "total": {
                          "type": "number",
                          "format": "decimal"
                        }
                      }
                    },
                    "out_trade_no": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Order with this out_trade_no already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/profile": {
      "get": {
        "tags": [
          "ProfileManagement"
        ],
        "summary": "Get current user's profile information",
        "description": "Retrieve authenticated user's profile data including personal information and preferences",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "User profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProfileManagement"
        ],
        "summary": "Update user profile information",
        "description": "Update authenticated user's profile details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "preferences": {
                    "type": "object",
                    "properties": {
                      "language": {
                        "type": "string",
                        "enum": [
                          "en",
                          "es",
                          "fr"
                        ]
                      },
                      "timezone": {
                        "type": "string"
                      },
                      "notification_email": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/miniprogram/profile": {
      "get": {
        "tags": ["Miniprogram"],
        "summary": "获取当前用户资料",
        "description": "获取已认证微信小程序用户的资料信息",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "用户资料获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniprogramUserProfile"
                }
              }
            }
          },
          "401": {
            "description": "未认证 - token 无效或缺失"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "post": {
        "tags": ["Miniprogram"],
        "summary": "更新用户资料",
        "description": "更新已认证用户的资料（支持部分更新）",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "description": "显示名称"
                  },
                  "nickname": {
                    "type": "string",
                    "maxLength": 50,
                    "description": "昵称（存储在 wechat_extra 中）"
                  }
                }
              },
              "example": {
                "name": "张三",
                "nickname": "小明"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "资料更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniprogramUserProfile"
                }
              }
            }
          },
          "400": {
            "description": "请求无效 - 未提供要更新的字段"
          },
          "401": {
            "description": "未认证 - token 无效或缺失"
          },
          "422": {
            "description": "参数校验失败"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MiniprogramUserProfile": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "description": "用户ID"
          },
          "name": {
            "type": "string",
            "description": "显示名称"
          },
          "nickname": {
            "type": "string",
            "nullable": true,
            "description": "昵称（来自 wechat_extra）"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "已绑定的手机号"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "最后更新时间"
          }
        },
        "required": ["user_id", "name", "updated_at"]
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "preferences": {
            "type": "object",
            "properties": {
              "language": {
                "type": "string"
              },
              "timezone": {
                "type": "string"
              },
              "notification_email": {
                "type": "boolean"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated endpoints"
      },
      "otaApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for OTA platform integration"
      }
    }
  },
...
}